Begins Excision of jQuery and Popper
Some cases of the wrong script being included. Other cases where a script (such as jQuery, Popper, and/or BootstrapJS) is referenced but never used. Where jQuery is used, a native alternative (Fetch) is utilized in its place. Because scripts are referenced at the bottom of the document, there is no need to bind to onLoad or listen for DOMContentLoaded either.

Addresses #16